More AutoCache improvements, Fixed duplicated records in SubmissionDataSource

+++ CoreData Models/CommentModel+CoreDataProperties.swift
@@ -20,10 +20,10 @@ extension CommentModel {
     @NSManaged public var approvedBy: String?
     @NSManaged public var author: String
     @NSManaged public var awardsJSON: String?
-    @NSManaged public var controversality: Int32
+    @NSManaged public var controversality: Int64
     @NSManaged public var created: Date
     @NSManaged public var edited: Date
-    @NSManaged public var depth: Int32
+    @NSManaged public var depth: Int64
     @NSManaged public var distinguished: String?
     @NSManaged public var flairJSON: String?
     @NSManaged public var hasVoted: Bool
@@ -48,7 +48,7 @@ extension CommentModel {
     @NSManaged public var removalReason: String?
     @NSManaged public var removedBy: String?
     @NSManaged public var reportsJSON: String?
-    @NSManaged public var score: Int32
+    @NSManaged public var score: Int64
     @NSManaged public var scoreHidden: Bool
     @NSManaged public var submissionTitle: String
     @NSManaged public var subreddit: String
