+++ Slide for Reddit/RedditObjects/MoreObject.swift
@@ -14,7 +14,7 @@ class MoreObject: RedditObject {
     public var id: String
     public var name: String
     public var parentID: String
-    public var count: Int32
+    public var count: Int64
     public var childrenString: String
     
     static func moreToMoreObject(more: More) -> MoreObject {
@@ -29,7 +29,7 @@ class MoreObject: RedditObject {
         }
         self.name = more.name
         self.parentID = more.parentId
-        self.count = Int32(more.count)
+        self.count = Int64(more.count)
         self.childrenString = more.children.joined()
     }
 }
